[C2X] N3007 Type inference for object definitions
Authored by to268 on Sep 5 2022, 4:30 AM.

This patches implements the auto keyword from the N3007 standard specification.
This allows deducing the type of the variable like in C++

auto nb = 1;
auto chr = 'A';
auto str = "String";

The list of statements which allows the usage of auto:

    Basic variables declarations (int, float, double, char, char*...)
    Macros declaring a variable with the auto type

The list of statements which will not work with the auto keyword:

    auto arrays
    sizeof(), alignas()
    auto parameters, auto return type
    auto as a struct/typedef member
    uninitialized auto variables
    auto in an union
    auto as a enum type specifier
    auto casts
    auto in an compound statement

I have updated and fixed all minors comments about the test cases.
I have added the diagnostic to warn users about using a Clang extension when declaring an explicit auto* in pedantic mode.
I need to debug more about the two remaining diagnostics to fix:

    Wrong diagnostic when using the _Atomic attribute
    Misleading diagnostic where it should say that auto is not allowed but says that [...] requires an initializer.

https://reviews.llvm.org/D133289